Separate os dependent syscalls into separate files
