+++ tools/cluster/cluster.go
@@ -12,7 +12,6 @@ import (
 	"path"
 	"sort"
 	"strconv"
-	"syscall"
 	"text/template"
 	"time"
 
@@ -342,30 +341,6 @@ func (clu *Cluster) KillNode(nodeIndex int) error {
 	}
 
 	return nil
-}
-
-func (clu *Cluster) FreezeNode(nodeIndex int) error {
-	if nodeIndex >= len(clu.waspCmds) {
-		return xerrors.Errorf("[cluster] Wasp node with index %d not found, active processes: %v", nodeIndex, len(clu.waspCmds))
-	}
-
-	process := clu.waspCmds[nodeIndex]
-
-	err := process.Process.Signal(syscall.SIGSTOP)
-
-	return err
-}
-
-func (clu *Cluster) UnfreezeNode(nodeIndex int) error {
-	if nodeIndex >= len(clu.waspCmds) {
-		return xerrors.Errorf("[cluster] Wasp node with index %d not found", nodeIndex)
-	}
-
-	process := clu.waspCmds[nodeIndex]
-
-	err := process.Process.Signal(syscall.SIGCONT)
-
-	return err
 }
 
 func (clu *Cluster) RestartNode(nodeIndex int) error {
